change cell structure
