+++ src/components/Table/Table.tsx
@@ -4,7 +4,6 @@ import classNames from 'classnames';
 import Body from './TableBody/TableBody';
 import Cell from './TableCell/TableCell';
 import Header from './TableHeader/TableHeader';
-import HeaderCell from './TableHeaderCell/TableHeaderCell';
 import Row from './TableRow/TableRow';
 
 import cssReset from '../../css-reset.module.css';
@@ -17,10 +16,9 @@ type Props = {
 
 const Table: React.FC<Props> & {
   Header: typeof Header;
-  HeaderCell: typeof HeaderCell;
   Body: typeof Body;
   Row: typeof Row;
-  Cell: typeof HeaderCell;
+  Cell: typeof Cell;
 } = ({ className, children }: Props) => {
   const tableClassNames = classNames(cssReset.ventura, styles.table, className);
 
@@ -28,7 +26,6 @@ const Table: React.FC<Props> & {
 };
 
 Table.Header = Header;
-Table.HeaderCell = HeaderCell;
 Table.Body = Body;
 Table.Row = Row;
 Table.Cell = Cell;
