+++ src/components/Table/TableCell/TableCell.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import classNames from 'classnames';
 
 import styles from './TableCell.module.css';
@@ -10,7 +10,15 @@ type Props = {
   align?: 'center' | 'left' | 'right';
 };
 
+export interface CellContext {
+  location: 'body' | 'header';
+}
+
+export const cellContext = React.createContext<CellContext>({ location: 'body' });
+
 const TableCell: React.FC<Props> = ({ className, children, colSpan, align = 'left' }: Props) => {
+  const cellContextValue = useContext(cellContext);
+
   const tableCellClassNames = classNames(
     {
       [styles.cellAlignLeft]: align === 'left',
@@ -18,10 +26,26 @@ const TableCell: React.FC<Props> = ({ className, children, colSpan, align = 'lef
       [styles.cellAlignCenter]: align === 'center',
     },
     styles.cell,
-    styles.row,
     className,
   );
 
+  const tableHeaderCellClassNames = classNames(
+    {
+      [styles.cellAlignLeft]: align === 'left',
+      [styles.cellAlignRight]: align === 'right',
+      [styles.cellAlignCenter]: align === 'center',
+    },
+    styles.headerCell,
+    className,
+  );
+
+  if (cellContextValue.location === 'header') {
+    return (
+      <th colSpan={colSpan} className={tableHeaderCellClassNames}>
+        {children}
+      </th>
+    );
+  }
   return (
     <td colSpan={colSpan} className={tableCellClassNames}>
       {children}
