+++ src/components/Table/TableHeader/TableHeader.tsx
@@ -1,14 +1,20 @@
 import React from 'react';
 
+import { CellContext, cellContext } from '../TableCell/TableCell';
+
 type Props = {
   className?: string;
   children?: React.ReactNode;
 };
 
 const TableHeader: React.FC<Props> = ({ className, children }: Props) => {
+  const cellContextValues: CellContext = { location: 'header' };
+
   return (
     <thead className={className}>
-      <tr>{children}</tr>
+      <cellContext.Provider value={cellContextValues}>
+        <tr>{children}</tr>
+      </cellContext.Provider>
     </thead>
   );
 };
