--- src/components/Table/TableHeaderCell/TableHeaderCell.module.css
@@ -1,29 +1,0 @@
-.headerCell {
-  background-color: var(--color-primary-hover);
-  height: 36px;
-  line-height: 36px;
-  padding-left: 14px;
-  padding-right: 6px;
-  color: var(--font-color-sub)
-}
-
-.cellAlignLeft {
-  text-align: left;
-}
-
-.cellAlignRight {
-  text-align: right;
-}
-
-.cellAlignCenter {
-  text-align: center;
-}
-
-
-.headerCell:first-child {
-  border-radius: var(--border-radius-small) 0 0 var(--border-radius-small);
-}
-
-.headerCell:last-child {
-  border-radius: 0 var(--border-radius-small) var(--border-radius-small) 0;
-}
